+++ src/fabricStudio/FabricStudioConfig.ts
@@ -1,0 +1,5 @@
+export class FabricStudioConfig {
+	static get FSScheme(): string {
+		return "fabric";
+	}
+}