+++ src/vscode/treeviews/workspaces/PowerBIDatasetTablePartitions.ts
@@ -13,6 +13,7 @@ import { PowerBIDatasetTable } from './PowerBIDatasetTable';
 import { PowerBIWorkspaceGenericFolder } from './PowerBIWorkspaceGenericFolder';
 import { PowerBIDatasetTableColumn } from './PowerBIDatasetTableColumn';
 import { PowerBIDatasetTablePartition } from './PowerBIDatasetTablePartition';
+import { PowerBIConfiguration } from '../../configuration/PowerBIConfiguration';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class PowerBIDatasetTablePartitions extends PowerBIWorkspaceGenericFolder {
@@ -36,12 +37,34 @@ export class PowerBIDatasetTablePartitions extends PowerBIWorkspaceGenericFolder
 			let children: PowerBIDatasetTablePartition[] = [];
 
 			try {
-				const items: iPowerBIDatasetDMV[] = await PowerBIApiService.getDMV(this.apiPath, "PARTITIONS", "[TableID] = " + this.table.tableId);
+				if (PowerBIConfiguration.useFabricStudio) {
+					const partitionRegex = /^\s*partition\s*(?<partitionName>[^\s]*)[\s=]/gm;
 
-				for (let item of items) {
-					let treeItem = new PowerBIDatasetTablePartition(item, this.groupId, this);
-					children.push(treeItem);
-					PowerBICommandBuilder.pushQuickPickItem(treeItem);
+					const tmdl = await this.table.getTMDLContent();
+					const matches = tmdl.matchAll(partitionRegex);
+    
+					for (const match of matches) {
+						const partitionName = match.groups.partitionName;
+
+						const meta: iPowerBIDatasetDMV = {
+							"name": partitionName,
+							"id": partitionName,
+							"properties": {}
+						};
+
+						let treeItem = new PowerBIDatasetTablePartition(meta, this.groupId, this);
+						children.push(treeItem);
+						PowerBICommandBuilder.pushQuickPickItem(treeItem);
+					}
+				}
+				else {
+					const items: iPowerBIDatasetDMV[] = await PowerBIApiService.getDMV(this.apiPath, "PARTITIONS", "[TableID] = " + this.table.tableId);
+
+					for (let item of items) {
+						let treeItem = new PowerBIDatasetTablePartition(item, this.groupId, this);
+						children.push(treeItem);
+						PowerBICommandBuilder.pushQuickPickItem(treeItem);
+					}
 				}
 			}
 			catch (e) {