+++ src/vscode/treeviews/workspaces/PowerBIDatasetTables.ts
@@ -10,6 +10,7 @@ import { PowerBICommandBuilder } from '../../../powerbi/CommandBuilder';
 import { ThisExtension } from '../../../ThisExtension';
 import { PowerBIDatasetTable } from './PowerBIDatasetTable';
 import { PowerBIWorkspaceGenericFolder } from './PowerBIWorkspaceGenericFolder';
+import { PowerBIConfiguration } from '../../configuration/PowerBIConfiguration';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class PowerBIDatasetTables extends PowerBIWorkspaceGenericFolder {
@@ -25,6 +26,12 @@ export class PowerBIDatasetTables extends PowerBIWorkspaceGenericFolder {
 		return this.parent as PowerBIDataset;
 	}
 
+	get fabricFsUri(): vscode.Uri {
+		const fabricUri = vscode.Uri.joinPath(this.dataset.fabricFsUri, "definition", "tables");
+
+		return fabricUri;
+	}
+
 	async getChildren(element?: PowerBIWorkspaceTreeItem): Promise<PowerBIWorkspaceTreeItem[]> {
 		if (element != null && element != undefined) {
 			return element.getChildren();
@@ -33,16 +40,43 @@ export class PowerBIDatasetTables extends PowerBIWorkspaceGenericFolder {
 			let children: PowerBIDatasetTable[] = [];
 
 			try {
-				const items: iPowerBIDatasetDMV[] = await PowerBIApiService.getDMV(this.apiPath, "TABLES");
 
-				for (let item of items) {
-					let treeItem = new PowerBIDatasetTable(item, this.groupId, this);
-					children.push(treeItem);
-					PowerBICommandBuilder.pushQuickPickItem(treeItem);
+				if (PowerBIConfiguration.useFabricStudio) {
+					await ThisExtension.ensureFabricStudio();
+					const fabricUri = this.dataset.fabricFsUri;
+
+					// we need to read the root folder first to ensure the child folders are loaded 
+					const initFolder = await vscode.workspace.fs.readDirectory(fabricUri);
+					const folders = await vscode.workspace.fs.readDirectory(this.fabricFsUri);
+
+					for (let item of folders) {
+						if (item[1] != vscode.FileType.File) {
+							ThisExtension.log("Item is not a TMDL File: Skipping " + item[0]);
+							continue;
+						}
+						const tableName = item[0].replace("\.tmdl", "")
+						const meta: iPowerBIDatasetDMV = {
+							"name": tableName,
+							"id": tableName,
+							"properties": {}
+						};
+						let treeItem = new PowerBIDatasetTable(meta, this.groupId, this);
+						children.push(treeItem);
+						PowerBICommandBuilder.pushQuickPickItem(treeItem);
+					}
+				}
+				else {
+					const items: iPowerBIDatasetDMV[] = await PowerBIApiService.getDMV(this.apiPath, "TABLES");
+
+					for (let item of items) {
+						let treeItem = new PowerBIDatasetTable(item, this.groupId, this);
+						children.push(treeItem);
+						PowerBICommandBuilder.pushQuickPickItem(treeItem);
+					}
 				}
 			}
 			catch (e) {
-				if(e.message.includes("PowerBIFeatureDisabled")) {
+				if (e.message.includes("PowerBIFeatureDisabled")) {
 					ThisExtension.log(e.message);
 					vscode.window.showErrorMessage(e.message);
 				}